ireps_scraper/requirements.txt
playwright
2captcha-python
flask
python-dotenv
requests
